Bug line 28: average should be double type.

+++ bugExample.cpp
@@ -25,7 +25,7 @@ void multArrays(const int[], const int[], int[], const int);
 void displayArray(const int[], const int);
 int sumOddArray(const int[], const int);
 bool isAllPositive(const int[], const int);
-bool avgOddArray(const int[], const int, int&);
+bool avgOddArray(const int[], const int, double&);
 void revertNumbers(int arr[], const int size);
 
 /**
